履约

+++ src/test/java/com/shopify/model/ShopifyProductCreationRequestTest.java
@@ -114,11 +114,11 @@ public class ShopifyProductCreationRequestTest {
 
 		final List<ShopifyVariant> actualShopifyVariants = actualShopifyProduct.getVariants();
 		assertEquals(3, actualShopifyVariants.size());
-		assertEquals(1, actualShopifyVariants.get(0).getPosition().intValue());
+		assertEquals(1, actualShopifyVariants.get(0).getPosition());
 		assertSame(secondVariantCreationRequest.getRequest(), actualShopifyVariants.get(0));
-		assertEquals(2, actualShopifyVariants.get(1).getPosition().intValue());
+		assertEquals(2, actualShopifyVariants.get(1).getPosition());
 		assertSame(thirdVariantCreationRequest.getRequest(), actualShopifyVariants.get(1));
-		assertEquals(3, actualShopifyVariants.get(2).getPosition().intValue());
+		assertEquals(3, actualShopifyVariants.get(2).getPosition());
 		assertSame(firstVariantCreationRequest.getRequest(), actualShopifyVariants.get(2));
 
 		assertFalse(actualShopifyProductCreationRequest.hasVariantImagePosition(1));
@@ -207,11 +207,11 @@ public class ShopifyProductCreationRequestTest {
 
 		final List<ShopifyVariant> actualShopifyVariants = actualShopifyProduct.getVariants();
 		assertEquals(3, actualShopifyVariants.size());
-		assertEquals(1, actualShopifyVariants.get(0).getPosition().intValue());
+		assertEquals(1, actualShopifyVariants.get(0).getPosition());
 		assertSame(secondVariantCreationRequest.getRequest(), actualShopifyVariants.get(0));
-		assertEquals(2, actualShopifyVariants.get(1).getPosition().intValue());
+		assertEquals(2, actualShopifyVariants.get(1).getPosition());
 		assertSame(thirdVariantCreationRequest.getRequest(), actualShopifyVariants.get(1));
-		assertEquals(3, actualShopifyVariants.get(2).getPosition().intValue());
+		assertEquals(3, actualShopifyVariants.get(2).getPosition());
 		assertSame(firstVariantCreationRequest.getRequest(), actualShopifyVariants.get(2));
 
 		assertFalse(actualShopifyProductCreationRequest.hasVariantImagePosition(1));
